Add 'includes' helper
Closes #6
Closes #8
