+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.1 (beta)
 
 - Added support for matching identity via `expect(actual).is_(expected)`.
+- Added `includes` helper as an alias of `contains`.
 
 # 2.0 (2020-02-06)
