+++ pyproject.toml
@@ -1,7 +1,7 @@
 [tool.poetry]
 
 name = "pytest-expecter"
-version = "2.1b1"
+version = "2.1b2"
 description = "Better testing with expecter and pytest."
 
 license = "BSD"
@@ -20,6 +20,7 @@ classifiers = [
     "Environment :: Plugins",
     "Framework :: Pytest",
     "Intended Audience :: Developers",
+    "License :: OSI Approved :: BSD License",
     "Operating System :: OS Independent",
     "Programming Language :: Python",
     "Programming Language :: Python :: 3.6",
